requirements.txt
customtkinter>=5.2.2
httpx>=0.27.0
holehe
sherlock-project
python-whois
dnspython
Pillow
PyPDF2
psutil
